Move parsing to separate file
